+++ UnitTests/BraintreeCoreTests/Helpers/FakeRequest.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct FakeRequest: Encodable {
+    
+    let testValue: String
+}